+++ src/Cake.CMake/Arguments/QuotedArgumentList.cs
@@ -1,0 +1,7 @@
+﻿namespace Cake.CMake.Arguments
+{
+  internal class QuotedArgumentList
+  {
+    
+  }
+}