+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0-beta01] - 2019-08-05
+### Added 
+- Support build mode for CMake.(--build flag)
+### Changed
+- Removed sourcePath from required parameters for CMake command.
 
 ## [1.0.0] - 2019-08-02
 ### Changed
@@ -20,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
  - Updated Cake.Core dependency.
 
-## [0.1.0] - 2016-04-10
+## [0.1.0] - 2015-04-10
 ### Added
 - Linux support.